requirements.txt
numpy
opencv-python
pillow
tqdm
scikit-image
pytorch-lightning
torch
torchvision
tensorflow